requirements.txt
wheel
tqdm
numpy
pandas
scipy
scikit_learn
matplotlib
scikit_image
Pillow
renumics-spotlight
ultralytics>=3.8.100
supervision==0.25.0
autodistill
roboflow

tator
panoptes-client